@@ -1,6 +1,7 @@
 # frozen_string_literal: true
+
 source "https://rubygems.org"
-ruby "2.3.1"
+ruby "2.5.0"
 
 # Database
 gem "pg"
@@ -10,46 +11,41 @@ gem "carrierwave"
 gem "cloudinary"
 
 # Assets
-gem "autoprefixer-rails", "~> 6.3.6"
+gem "autoprefixer-rails"
 gem "bootstrap-sass"
+gem "coffee-rails"
 gem "font-awesome-sass"
-gem "coffee-rails", "~> 4.1.0"
 gem "jquery-rails"
 gem "sass-rails"
-gem "uglifier", "~> 3.0.0"
-gem "sprockets", "~> 3.0"
+gem "uglifier"
 
 # Application
-gem "rails", "~> 4.2"
+gem "devise"
+gem "friendly_id"
 gem "omniauth-github"
 gem "omniauth-twitter"
-gem "friendly_id", "~> 5.1.0"
-gem "devise", "~> 3.5.2"
+gem "rails", "~> 5.1"
 
 # Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem "turbolinks"
-gem "jquery-turbolinks"
 gem "bourbon"
+gem "jquery-turbolinks"
+gem "turbolinks"
 
 # View
-gem "simple_form"
 gem "enumerize"
-gem "redcarpet"
 gem "gemoji"
+gem "redcarpet"
+gem "simple_form"
 
 # Server
 gem "unicorn"
-gem "newrelic_rpm", group: :production
-
-# Heroku Assets
-gem "rails_12factor", group: :production
 
 # Globalization/Localization
-gem "globalize"
+gem "globalize", "~> 5.1"
 gem "globalize-accessors"
 
 # Administration
-gem "administrate", "~> 0.1.5"
+gem "administrate"
 
 # Ops
 gem "bugsnag"
@@ -62,40 +58,39 @@ gem "twitter"
 # Utility
 gem "colorize"
 
+# Authorization system
+gem "pundit"
+
 group :development do
   gem "annotate"
-  gem "quiet_assets"
+  gem "bullet"
   gem "spring"
   gem "spring-commands-rspec"
-  gem "bullet"
-end
-
-group :legacy do
-  gem "mysql2"
-  gem "nokogiri"
 end
 
 group :development, :test do
+  gem "byebug"
   gem "capybara"
-  gem "factory_girl_rails"
+  gem "dotenv-rails"
+  gem "factory_bot_rails"
   gem "faker"
   gem "guard-ctags-bundler"
   gem "guard-rspec", require: false
   gem "guard-rubocop"
   gem "pry"
-  gem "byebug"
-  gem "dotenv-rails"
-  gem "rspec-rails"
-  gem "shoulda-matchers"
   gem "rspec-its"
-  gem "rubocop", "~> 0.45.0", require: false
+  gem "rspec-rails"
+  gem "rubocop", require: false
+  gem "shoulda-matchers"
   gem "sqlite3"
   gem "timecop"
 end
 
 group :test do
+  gem "rails-controller-testing" # this is temporary
   gem "vcr"
-  gem "webmock", ">= 1.8.0"
+  gem "webmock"
+  gem 'database_cleaner'
+  gem 'simplecov', require: false
 end
 
-gem "coveralls", require: false, group: :test
