@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-FactoryGirl.define do
+
+FactoryBot.define do
   factory :job do
     association :organization
     association :author, factory: :user
@@ -18,6 +19,11 @@ FactoryGirl.define do
 
     trait :archived do
       state :archived
+    end
+
+    trait :published_without_published_at do
+      state :published
+      after(:create) { |job| job.update(published_at: nil) }
     end
   end
 end
