@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-FactoryGirl.define do
+
+FactoryBot.define do
   factory :location do
     name { Faker::Company.name }
     address { Faker::Address.street_address + Faker::Address.city + Faker::Address.country }
