@@ -1,11 +1,12 @@
 # frozen_string_literal: true
+
 class HomeController < ApplicationController
   def index
     @past_events, @future_events = Event.published.includes(
       :translations,
       :location,
-      :talks,
-      :sponsors
-    ).order(:starts_at).limit(3).partition { |e| e.starts_at < Time.now }
+      :sponsors,
+      talks: %i(member translations)
+    ).order(:starts_at).limit(5).partition { |e| e.starts_at < Time.now }
   end
 end
