@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "administrate/base_dashboard"
 
 class JobDashboard < Administrate::BaseDashboard
@@ -13,9 +14,11 @@ class JobDashboard < Administrate::BaseDashboard
     title: Field::String,
     state: EnumField,
     organization: Field::BelongsTo,
+    author: Field::BelongsTo.with_options(class_name: "User"),
     description: MarkdownField,
     created_at: Field::DateTime,
-    updated_at: Field::DateTime
+    updated_at: Field::DateTime,
+    published_at: Field::DateTime
   }.freeze
 
   # COLLECTION_ATTRIBUTES
@@ -23,12 +26,13 @@ class JobDashboard < Administrate::BaseDashboard
   #
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
-  COLLECTION_ATTRIBUTES = [
-    :id,
-    :state,
-    :title,
-    :organization,
-  ].freeze
+  COLLECTION_ATTRIBUTES = %i(
+    id
+    state
+    title
+    organization
+    published_at
+  ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
@@ -37,12 +41,12 @@ class JobDashboard < Administrate::BaseDashboard
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
-  FORM_ATTRIBUTES = [
-    :state,
-    :title,
-    :organization,
-    :description
-  ].freeze
+  FORM_ATTRIBUTES = %i(
+    state
+    title
+    organization
+    description
+  ).freeze
 
   # Overwrite this method to customize how locations are displayed
   # across all pages of the admin dashboard.
