@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "administrate/base_dashboard"
 
 class MeetupDashboard < Administrate::BaseDashboard
@@ -23,12 +24,12 @@ class MeetupDashboard < Administrate::BaseDashboard
   #
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
-  COLLECTION_ATTRIBUTES = [
-    :location,
-    :translations,
-    :id,
-    :type
-  ].freeze
+  COLLECTION_ATTRIBUTES = %i(
+    location
+    translations
+    id
+    type
+  ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
@@ -37,12 +38,12 @@ class MeetupDashboard < Administrate::BaseDashboard
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
-  FORM_ATTRIBUTES = [
-    :location,
-    :translations,
-    :type,
-    :starts_at
-  ].freeze
+  FORM_ATTRIBUTES = %i(
+    location
+    translations
+    type
+    starts_at
+  ).freeze
 
   # Overwrite this method to customize how meetups are displayed
   # across all pages of the admin dashboard.
