@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "administrate/base_dashboard"
 
 class NewsItemDashboard < Administrate::BaseDashboard
@@ -25,11 +26,11 @@ class NewsItemDashboard < Administrate::BaseDashboard
   #
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
-  COLLECTION_ATTRIBUTES = [
-    :title,
-    :state,
-    :published_at
-  ].freeze
+  COLLECTION_ATTRIBUTES = %i(
+    title
+    state
+    published_at
+  ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
@@ -38,12 +39,12 @@ class NewsItemDashboard < Administrate::BaseDashboard
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
-  FORM_ATTRIBUTES = [
-    :title,
-    :body,
-    :state,
-    :published_at
-  ].freeze
+  FORM_ATTRIBUTES = %i(
+    title
+    body
+    state
+    published_at
+  ).freeze
 
   # Overwrite this method to customize how news items are displayed
   # across all pages of the admin dashboard.
