@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "administrate/base_dashboard"
 
 class OrganizationDashboard < Administrate::BaseDashboard
@@ -24,13 +25,13 @@ class OrganizationDashboard < Administrate::BaseDashboard
   #
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
-  COLLECTION_ATTRIBUTES = [
-    :id,
-    :logo,
-    :name,
-    :address,
-    :sponsored_events,
-  ].freeze
+  COLLECTION_ATTRIBUTES = %i(
+    id
+    logo
+    name
+    address
+    sponsored_events
+  ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
@@ -39,12 +40,12 @@ class OrganizationDashboard < Administrate::BaseDashboard
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
-  FORM_ATTRIBUTES = [
-    :name,
-    :address,
-    :logo,
-    :description
-  ].freeze
+  FORM_ATTRIBUTES = %i(
+    name
+    address
+    logo
+    description
+  ).freeze
 
   # Overwrite this method to customize how organizations are displayed
   # across all pages of the admin dashboard.
