@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "administrate/base_dashboard"
 
 class TalkDashboard < Administrate::BaseDashboard
@@ -18,6 +19,7 @@ class TalkDashboard < Administrate::BaseDashboard
     level: EnumField,
     format: EnumField,
     slides_url: Field::String,
+    youtube_id: Field::String,
     created_at: Field::DateTime,
     updated_at: Field::DateTime
   }.freeze
@@ -27,15 +29,14 @@ class TalkDashboard < Administrate::BaseDashboard
   #
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
-  COLLECTION_ATTRIBUTES = [
-    :id,
-    :title,
-    :state,
-    :format,
-    :event,
-    :member,
-    :title
-  ].freeze
+  COLLECTION_ATTRIBUTES = %i(
+    id
+    title
+    state
+    format
+    event
+    member
+  ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
@@ -44,16 +45,17 @@ class TalkDashboard < Administrate::BaseDashboard
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
-  FORM_ATTRIBUTES = [
-    :title,
-    :description,
-    :event,
-    :member,
-    :state,
-    :level,
-    :format,
-    :slides_url,
-  ].freeze
+  FORM_ATTRIBUTES = %i(
+    title
+    description
+    event
+    member
+    state
+    level
+    format
+    slides_url
+    youtube_id
+  ).freeze
 
   # Overwrite this method to customize how talks are displayed
   # across all pages of the admin dashboard.
