@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "administrate/base_dashboard"
 
 class UserDashboard < Administrate::BaseDashboard
@@ -20,12 +21,12 @@ class UserDashboard < Administrate::BaseDashboard
   #
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
-  COLLECTION_ATTRIBUTES = [
-    :id,
-    :email,
-    :member,
-    :admin,
-  ].freeze
+  COLLECTION_ATTRIBUTES = %i(
+    id
+    email
+    member
+    admin
+  ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
@@ -34,10 +35,10 @@ class UserDashboard < Administrate::BaseDashboard
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
-  FORM_ATTRIBUTES = [
-    :email,
-    :admin,
-  ].freeze
+  FORM_ATTRIBUTES = %i(
+    email
+    admin
+  ).freeze
 
   # Overwrite this method to customize how users are displayed
   # across all pages of the admin dashboard.
