@@ -1,12 +1,17 @@
 # frozen_string_literal: true
+
 class Job < ActiveRecord::Base
   extend Enumerize
+  include Authorable
   STATES = %w(draft published archived).freeze
 
+  before_save :set_published_at, if: proc { state_changed?(to: "published") }
+
   belongs_to :organization
-  belongs_to :author, foreign_key: :user_id, class_name: "User"
 
   scope :published, -> { where(state: :published).order(created_at: :desc) }
+  scope :search, -> (query = nil) { where("LOWER(title) LIKE LOWER(?) OR LOWER(description) LIKE LOWER(?)", "%#{query}%", "%#{query}%") if query }
+  scope :draft, -> { where(state: :draft) }
 
   enumerize :state, in: STATES, default: :draft
 
@@ -19,5 +24,13 @@ class Job < ActiveRecord::Base
   validates :state,
             presence: true,
             inclusion: { in: STATES }
-  validates :author, presence: true
+
+  accepts_nested_attributes_for :organization
+
+  private
+
+  def set_published_at
+    # We can't use touch since it only work on persisted object
+    self.published_at = DateTime.current
+  end
 end
