@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 # a NewsItem represents any announcement that we want to communicate to our
 # web audience.
 #
@@ -11,7 +12,7 @@
 #   can still resolve: http://www.montrealrb.com/[post_date:YYYY]/[post_date:MM]/[post_name]
 #
 class NewsItem < ActiveRecord::Base
-  belongs_to :author, foreign_key: :user_id, class_name: "User"
+  include Authorable
 
   # Extends
   extend Enumerize
@@ -37,7 +38,6 @@ class NewsItem < ActiveRecord::Base
   validates :state,
             presence: true,
             inclusion: { in: STATES }
-  validates :author, presence: true
 
   # Class methods
   enumerize :state, in: STATES, default: :draft
