@@ -1,12 +1,17 @@
 # frozen_string_literal: true
+
 class Talk < ActiveRecord::Base
   extend Enumerize
-  STATES  = [:proposed, :scheduled, :presented].freeze
-  FORMATS = [:full, :half, :lightning].freeze
-  LEVELS  = [:beginner, :intermediate, :advanced].freeze
+  STATES  = %i(proposed scheduled presented).freeze
+  FORMATS = %i(full half lightning).freeze
+  LEVELS  = %i(beginner intermediate advanced).freeze
 
   translates :title, :description
-  belongs_to :event, counter_cache: true
+  # Someday, we should be able to remove those lines
+  attribute :title
+  attribute :description
+
+  belongs_to :event, counter_cache: true, inverse_of: :talks
   belongs_to :member
 
   validates :title,  presence: true
