@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require File.expand_path("../boot", __FILE__)
 
 # Pick the frameworks you want:
@@ -21,7 +22,6 @@ module MontrealRb
     config.i18n.default_locale = :en
     config.i18n.available_locales = %i(en fr)
     config.i18n.enforce_available_locales = true
-    config.active_record.raise_in_transactional_callbacks = true
     config.to_prepare do
       Administrate::ApplicationController.helper MontrealRb::Application.helpers
     end
