@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-# Set up gems listed in the Gemfile.
+
 ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../../Gemfile", __FILE__)
 
-require "bundler/setup" if File.exist?(ENV["BUNDLE_GEMFILE"])
+require "bundler/setup" # Set up gems listed in the Gemfile.
