@@ -1,8 +1,9 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe Admin::EventsController, type: :controller do
-  let(:admin) { FactoryGirl.create(:user, :admin) }
+  let(:admin) { create(:user, :admin) }
   let(:valid_attributes) do
     attributes_for(:event).merge(author: nil,
                                  location_id: create(:location).id)
@@ -27,29 +28,29 @@ RSpec.describe Admin::EventsController, type: :controller do
 
     it "creates a new event" do
       expect do
-        post :create, event: valid_attributes
+        post :create, params: { event: valid_attributes }
       end.to change(Event, :count).by(1)
     end
 
     it "assigns the current user to the 'author' field" do
-      post :create, event: valid_attributes
+      post :create, params: { event: valid_attributes }
       expect(Event.last.author).to eq admin
     end
   end
 
   describe "PUT #update" do
-    let!(:event) { FactoryGirl.create(:event) }
+    let!(:event) { create(:event) }
     before do
       login_user(admin)
     end
 
     it "assigns the current user to the 'author' field" do
-      put :update, id: event.id, event: event.attributes
+      put :update, params: { id: event.id, event: event.attributes }
       expect(event.reload.author).to eq admin
     end
 
     context "the event has a tweet value set to true" do
-      let!(:event) { FactoryGirl.create(:event) }
+      let!(:event) { create(:event) }
       let(:valid_attributes) do
         attributes_for(:event).merge(tweet: true)
       end
@@ -57,13 +58,13 @@ RSpec.describe Admin::EventsController, type: :controller do
       it "tweets" do
         allow(controller).to receive(:requested_resource) { event }
         expect(event).to receive(:tweet)
-        put :update, id: event.id, event: valid_attributes
+        put :update, params: { id: event.id, event: valid_attributes }
       end
     end
   end
 
   describe "DELETE #destroy" do
-    let!(:event) { FactoryGirl.create(:event) }
+    let!(:event) { create(:event) }
     before do
       login_user(admin)
     end
@@ -71,9 +72,20 @@ RSpec.describe Admin::EventsController, type: :controller do
     context "success" do
       it "deletes an event" do
         expect do
-          delete :destroy, id: event.id
+          delete :destroy, params: { id: event.id }
         end.to change(Event, :count).by(-1)
         expect(response).to redirect_to admin_events_path
+      end
+    end
+
+    context "when cannot unlink all the talks" do
+      it "displays a flag message" do
+        allow(Event).to receive(:find).and_return(event)
+        expect(event).to receive(:unlink_all_talks).and_return(false)
+
+        delete :destroy, params: { id: event.id }
+
+        expect(flash[:error]).to eq("Could not delete event ##{event.id}")
       end
     end
   end
