@@ -1,8 +1,9 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe Admin::NewsItemsController, type: :controller do
-  let(:admin) { FactoryGirl.create(:user, :admin) }
+  let(:admin) { create(:user, :admin) }
   let(:valid_attributes) do
     attributes_for(:news_item).merge(author: nil)
   end
@@ -14,12 +15,12 @@ RSpec.describe Admin::NewsItemsController, type: :controller do
 
     it "creates a new news_item" do
       expect do
-        post :create, news_item: valid_attributes
+        post :create, params: { news_item: valid_attributes }
       end.to change(NewsItem, :count).by(1)
     end
 
     it "assigns the current user to the 'author' field" do
-      post :create, news_item: valid_attributes
+      post :create, params: { news_item: valid_attributes }
       expect(NewsItem.last.author).to eq admin
     end
   end
@@ -31,7 +32,7 @@ RSpec.describe Admin::NewsItemsController, type: :controller do
     end
 
     it "assigns the current user to the 'author' field" do
-      put :update, id: news_item.id, news_item: news_item.attributes
+      put :update, params: { id: news_item.id, news_item: news_item.attributes }
       expect(news_item.reload.author).to eq admin
     end
   end
