@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe Users::RegistrationsController do
@@ -15,7 +16,7 @@ RSpec.describe Users::RegistrationsController do
     end
 
     it "sends the slack invitation" do
-      post :create, user: attributes
+      post :create, params: { user: attributes }
       expect(service).to have_received(:call)
     end
   end
