@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe Organization, type: :model do
@@ -7,7 +8,7 @@ RSpec.describe Organization, type: :model do
   end
 
   describe "validation" do
-    subject { FactoryGirl.create(:organization) }
+    subject { create(:organization) }
 
     it { is_expected.to validate_presence_of(:name) }
     it { is_expected.to validate_uniqueness_of(:name) }
