@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe TweetEventService do
@@ -6,11 +7,14 @@ RSpec.describe TweetEventService do
   let(:service) { described_class.new(event) }
 
   subject { service.tap(&:call) }
+  before {
+    allow_any_instance_of(Twitter::REST::Client).to receive(:update).with(String)
+  }
 
   its(:success?) { is_expected.to eq true }
 
   it "tweets" do
-    allow_any_instance_of(Twitter::REST::Client).to receive(:update).with(String)
+    expect_any_instance_of(Twitter::REST::Client).to receive(:update).with(String)
     subject
   end
 
